Changed time no kill to quint32

+++ PredatorSimulator/PredatorSimulator/Attributes.h
@@ -18,7 +18,7 @@ struct RunnerAttributes : public Attributes
 
 struct PredatorAttributes : public Attributes
 {
-	quint8 timeNoKill;
+	quint32 timeNoKill;
 	qreal size;
 };
